add time line and Instant View

+++ app/src/main/java/com/lovejjfg/readhub/utils/UIUtil.java
@@ -143,7 +143,7 @@ public class UIUtil {
     }
 
 
-    public static int dip2px(Context context, double dpValue) {
+    public static int dip2px(Context context, float dpValue) {
         float density = context.getResources().getDisplayMetrics().density;
         return (int) (dpValue * density + 0.5);
     }
